Recycle log file (#886)
### What problem does this PR solve?

1. Add: recycle log file after checkpoint. recycle wal log after both
checkpoint and recycle delta checkpoint after full checkpoint.
2. Refactor: parse process of log filename.
3. Fix: cleanup bug after replay. #727.
 Cleanup may encounter not empty directory under such circumstance:
 3.1. Create table, add some block. Write delta catalog file.
 3.2. Drop table and write another delta catalog file.
3.3. Replay delta catalog files. "add some block" ops will be pruned in
replay stage. so not appears in catalog tree.
3.4. When cleanup, the table directory is not empty. because block
cannot be found in catalog.
 3.5. Solve: Cleanup will remove the whole directory.
6. Fix: force(manual) checkpoint will recycle log file now.
7. Fix: checkpoint should always write wal. (even if checkpoint file is
empty)
8. Add: unit test for recycle.

TODO:
record all checkpoint file path in the last checkpoint file.

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [x] New Feature (non-breaking change which adds functionality)
- [x] Test cases

+++ src/storage/meta/entry/table_index_entry.cpp
@@ -398,8 +398,13 @@ void TableIndexEntry::Cleanup() {
     }
 
     LOG_INFO(fmt::format("Cleanup dir: {}", *index_dir_));
+
+    // FIXME(sys): delete full text index by whole directory tmply, should call CleanupScanner::CleanupDir
     LocalFileSystem fs;
-    fs.DeleteDirectory(*index_dir_); // FIXME(sys): delete full text index by whole directory tmply
+    if (!fs.Exists(*index_dir_)) {
+        return;
+    }
+    fs.DeleteDirectory(*index_dir_);
 }
 
 void TableIndexEntry::PickCleanup(CleanupScanner *scanner) {}
